+++ Bee/ViewModels/DonateViewModel.cs
@@ -1,0 +1,8 @@
+
+using Bee.Base.ViewModels;
+
+namespace Bee.ViewModels;
+
+public partial class DonateViewModel : PageViewModelBase
+{
+}