+++ Bee/Views/DonateView/DonateView.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace Bee.Views;
+
+public partial class DonateView : UserControl
+{
+    public DonateView()
+    {
+        InitializeComponent();
+    }
+}